refactored OnBoarding and SchedulingParameters code and made separate components, added useParam router
